Enabled travis CI support and fixed compiler warnings
Travis CI will build RBDL once with SimpleMath support and once with
Eigen3. The benchmark addon is included to validate performance on
Travis CI infrastructure. The builds use either GCC or clang to compile.
The pre_script performs configuration of the build via cmake defines.
The test script itself 1) compiles RBDL with "-Wall -Wpedantic -Wextra
-Werror", 2) runs the unit tests, and 3) runs the benchmark suite.

As part of the work, all warnings for RBDLi, benchmark and unittests were fixed.
Additionally the cmake infrastructure required small adaptions to build
everyting with c++11 support, i.e., rbdk, tests, benchmarks.

+++ src/Kinematics.cc
@@ -52,8 +52,8 @@ RBDL_DLLAPI void UpdateKinematics(
       if (model.mJoints[i].mDoFCount == 1) {
         model.a[i] = model.a[i] + model.S[i] * QDDot[q_index];
       } else if (model.mJoints[i].mDoFCount == 3) {
-        Vector3d omegadot_temp (QDDot[q_index], 
-            QDDot[q_index + 1], 
+        Vector3d omegadot_temp (QDDot[q_index],
+            QDDot[q_index + 1],
             QDDot[q_index + 2]);
         model.a[i] = model.a[i] + model.multdof3_S[i] * omegadot_temp;
       }
@@ -63,7 +63,7 @@ RBDL_DLLAPI void UpdateKinematics(
       unsigned int joint_dof_count = custom_joint->mDoFCount;
 
       model.a[i] = model.a[i]
-        + ( model.mCustomJoints[custom_index]->S 
+        + ( model.mCustomJoints[custom_index]->S
             * QDDot.block(q_index, 0, joint_dof_count, 1));
     }
   }
@@ -132,10 +132,10 @@ RBDL_DLLAPI void UpdateKinematicsCustom(
         if (model.mJoints[i].mDoFCount == 1) {
           model.a[i] = model.a[i] + model.S[i] * (*QDDot)[q_index];
         } else if (model.mJoints[i].mDoFCount == 3) {
-          Vector3d omegadot_temp ((*QDDot)[q_index], 
-              (*QDDot)[q_index + 1], 
+          Vector3d omegadot_temp ((*QDDot)[q_index],
+              (*QDDot)[q_index + 1],
               (*QDDot)[q_index + 2]);
-          model.a[i] = model.a[i] 
+          model.a[i] = model.a[i]
             + model.multdof3_S[i] * omegadot_temp;
         }
       } else {
@@ -148,7 +148,7 @@ RBDL_DLLAPI void UpdateKinematicsCustom(
           + (  (model.mCustomJoints[k]->S)
               *(QDDot->block(q_index, 0, joint_dof_count, 1)));
       }
-    } 
+    }
   }
 }
 
@@ -167,15 +167,15 @@ RBDL_DLLAPI Vector3d CalcBodyToBaseCoordinates (
     unsigned int fbody_id = body_id - model.fixed_body_discriminator;
     unsigned int parent_id = model.mFixedBodies[fbody_id].mMovableParent;
 
-    Matrix3d fixed_rotation = 
+    Matrix3d fixed_rotation =
       model.mFixedBodies[fbody_id].mParentTransform.E.transpose();
     Vector3d fixed_position = model.mFixedBodies[fbody_id].mParentTransform.r;
 
     Matrix3d parent_body_rotation = model.X_base[parent_id].E.transpose();
     Vector3d parent_body_position = model.X_base[parent_id].r;
 
-    return (parent_body_position 
-        + (parent_body_rotation 
+    return (parent_body_position
+        + (parent_body_rotation
           * (fixed_position + fixed_rotation * (point_body_coordinates))) );
   }
 
@@ -205,9 +205,9 @@ RBDL_DLLAPI Vector3d CalcBaseToBodyCoordinates (
     Matrix3d parent_body_rotation = model.X_base[parent_id].E;
     Vector3d parent_body_position = model.X_base[parent_id].r;
 
-    return (fixed_rotation 
-        * ( - fixed_position 
-          - parent_body_rotation 
+    return (fixed_rotation
+        * ( - fixed_position
+          - parent_body_rotation
           * (parent_body_position - point_base_coordinates)));
   }
 
@@ -229,8 +229,8 @@ RBDL_DLLAPI Matrix3d CalcBodyWorldOrientation(
 
   if (body_id >= model.fixed_body_discriminator) {
     unsigned int fbody_id = body_id - model.fixed_body_discriminator;
-    model.mFixedBodies[fbody_id].mBaseTransform = 
-      model.mFixedBodies[fbody_id].mParentTransform 
+    model.mFixedBodies[fbody_id].mBaseTransform =
+      model.mFixedBodies[fbody_id].mParentTransform
       * model.X_base[model.mFixedBodies[fbody_id].mMovableParent];
 
     return model.mFixedBodies[fbody_id].mBaseTransform.E;
@@ -253,12 +253,12 @@ RBDL_DLLAPI void CalcPointJacobian (
     UpdateKinematicsCustom (model, &Q, NULL, NULL);
   }
 
-  SpatialTransform point_trans = 
-    SpatialTransform (Matrix3d::Identity(), 
-        CalcBodyToBaseCoordinates ( model, 
-          Q, 
+  SpatialTransform point_trans =
+    SpatialTransform (Matrix3d::Identity(),
+        CalcBodyToBaseCoordinates ( model,
+          Q,
           body_id,
-          point_position, 
+          point_position,
           false));
 
   assert (G.rows() == 3 && G.cols() == model.qdot_size );
@@ -295,7 +295,7 @@ RBDL_DLLAPI void CalcPointJacobian (
       G.block(0, q_index, 3, model.mCustomJoints[k]->mDoFCount) =
         ((point_trans
           * model.X_base[j].inverse()).toMatrix()
-         * model.mCustomJoints[k]->S).block( 
+         * model.mCustomJoints[k]->S).block(
            3,0,3,model.mCustomJoints[k]->mDoFCount);
     }
 
@@ -455,15 +455,15 @@ RBDL_DLLAPI Vector3d CalcPointVelocity (
   if (model.IsFixedBodyId(body_id)) {
     unsigned int fbody_id = body_id - model.fixed_body_discriminator;
     reference_body_id = model.mFixedBodies[fbody_id].mMovableParent;
-    Vector3d base_coords = 
+    Vector3d base_coords =
       CalcBodyToBaseCoordinates(model, Q, body_id, point_position, false);
     reference_point =
       CalcBaseToBodyCoordinates(model, Q, reference_body_id, base_coords,false);
   }
 
-  SpatialVector point_spatial_velocity = 
+  SpatialVector point_spatial_velocity =
     SpatialTransform (
-        CalcBodyWorldOrientation (model, Q, reference_body_id, false).transpose(), 
+        CalcBodyWorldOrientation (model, Q, reference_body_id, false).transpose(),
         reference_point).apply(model.v[reference_body_id]);
 
   return Vector3d (
@@ -499,14 +499,14 @@ RBDL_DLLAPI Math::SpatialVector CalcPointVelocity6D(
   if (model.IsFixedBodyId(body_id)) {
     unsigned int fbody_id = body_id - model.fixed_body_discriminator;
     reference_body_id = model.mFixedBodies[fbody_id].mMovableParent;
-    Vector3d base_coords = 
+    Vector3d base_coords =
       CalcBodyToBaseCoordinates(model, Q, body_id, point_position, false);
-    reference_point = 
+    reference_point =
       CalcBaseToBodyCoordinates(model, Q, reference_body_id, base_coords,false);
   }
 
   return SpatialTransform (
-      CalcBodyWorldOrientation (model, Q, reference_body_id, false).transpose(), 
+      CalcBodyWorldOrientation (model, Q, reference_body_id, false).transpose(),
       reference_point).apply(model.v[reference_body_id]);
 }
 
@@ -535,9 +535,9 @@ RBDL_DLLAPI Vector3d CalcPointAcceleration (
   if (model.IsFixedBodyId(body_id)) {
     unsigned int fbody_id = body_id - model.fixed_body_discriminator;
     reference_body_id = model.mFixedBodies[fbody_id].mMovableParent;
-    Vector3d base_coords = 
+    Vector3d base_coords =
       CalcBodyToBaseCoordinates (model, Q, body_id, point_position, false);
-    reference_point = 
+    reference_point =
       CalcBaseToBodyCoordinates (model, Q, reference_body_id,base_coords,false);
   }
 
@@ -582,9 +582,9 @@ RBDL_DLLAPI SpatialVector CalcPointAcceleration6D(
   if (model.IsFixedBodyId(body_id)) {
     unsigned int fbody_id = body_id - model.fixed_body_discriminator;
     reference_body_id = model.mFixedBodies[fbody_id].mMovableParent;
-    Vector3d base_coords = 
+    Vector3d base_coords =
       CalcBodyToBaseCoordinates (model, Q, body_id, point_position, false);
-    reference_point = 
+    reference_point =
       CalcBaseToBodyCoordinates (model, Q, reference_body_id,base_coords,false);
   }
 
@@ -595,7 +595,7 @@ RBDL_DLLAPI SpatialVector CalcPointAcceleration6D(
   SpatialVector p_v_i = p_X_i.apply(model.v[reference_body_id]);
   Vector3d a_dash = Vector3d (p_v_i[0], p_v_i[1], p_v_i[2]
       ).cross(Vector3d (p_v_i[3], p_v_i[4], p_v_i[5]));
-  return (p_X_i.apply(model.a[reference_body_id]) 
+  return (p_X_i.apply(model.a[reference_body_id])
       + SpatialVector (0, 0, 0, a_dash[0], a_dash[1], a_dash[2]));
 }
 
@@ -624,14 +624,14 @@ RBDL_DLLAPI bool InverseKinematics (
     for (unsigned int k = 0; k < body_id.size(); k++) {
       MatrixNd G (MatrixNd::Zero(3, model.qdot_size));
       CalcPointJacobian (model, Qres, body_id[k], body_point[k], G, false);
-      Vector3d point_base = 
+      Vector3d point_base =
         CalcBodyToBaseCoordinates (model, Qres, body_id[k], body_point[k], false);
       LOG << "current_pos = " << point_base.transpose() << std::endl;
 
       for (unsigned int i = 0; i < 3; i++) {
         for (unsigned int j = 0; j < model.qdot_size; j++) {
           unsigned int row = k * 3 + i;
-          LOG << "i = " << i << " j = " << j << " k = " << k << " row = " 
+          LOG << "i = " << i << " j = " << j << " k = " << k << " row = "
             << row << " col = " << j << std::endl;
           J(row, j) = G (i,j);
         }
@@ -649,8 +649,8 @@ RBDL_DLLAPI bool InverseKinematics (
       return true;
     }
 
-    MatrixNd JJTe_lambda2_I = 
-      J * J.transpose() 
+    MatrixNd JJTe_lambda2_I =
+      J * J.transpose()
       + lambda*lambda * MatrixNd::Identity(e.size(), e.size());
 
     VectorNd z (body_id.size() * 3);
@@ -708,7 +708,7 @@ Vector3d CalcAngularVelocityfromMatrix (
     double PI = atan(1)*4.0;
     return Vector3d (PI/2*(RotMat(0,0) + 1.0),PI/2*(RotMat(1,1) + 1.0),PI/2*(RotMat(2,2) + 1.0));
   }
-} 
+}
 
 RBDL_DLLAPI
 InverseKinematicsConstraintSet::InverseKinematicsConstraintSet() {
@@ -855,16 +855,16 @@ bool InverseKinematics (
 
     //     // "task space" from puppeteer
     //     MatrixNd Ek = MatrixNd::Zero (CS.e.size(), CS.e.size());
-    // 
+    //
     //     for (size_t ei = 0; ei < CS.e.size(); ei ++) {
     // //      Ek(ei,ei) = CS.error_norm * CS.error_norm * 0.5 + CS.lambda;
     //       Ek(ei,ei) = CS.e[ei]*CS.e[ei] * 0.5 + CS.lambda;
     //     }
-    // 
+    //
     //     MatrixNd JJT_Ek_wnI = CS.J * CS.J.transpose() + Ek;
-    // 
+    //
     //     VectorNd delta_theta = CS.J.transpose() * JJT_Ek_wnI.colPivHouseholderQr().solve (CS.e);
-    // 
+    //
     //     LOG << "change = " << delta_theta << std::endl;
 
 
@@ -872,7 +872,7 @@ bool InverseKinematics (
 
     double Ek = 0.;
 
-    for (size_t ei = 0; ei < CS.e.size(); ei ++) {
+    for (unsigned int ei = 0; ei < CS.e.size(); ei ++) {
       Ek += CS.e[ei] * CS.e[ei] * 0.5;
     }
 
@@ -881,7 +881,7 @@ bool InverseKinematics (
 
     assert (ek.size() == Qres.size());
 
-    for (size_t wi = 0; wi < Qres.size(); wi++) {
+    for (unsigned int wi = 0; wi < Qres.size(); wi++) {
       Wn(wi, wi) = ek[wi] * ek[wi] * 0.5 + CS.lambda;
       //      Wn(wi, wi) = Ek + 1.0e-3;
     }
